added a label for the meta section of the new child dialogue... and made table sort ascending
